ovis/requirements.txt
PyQt6==6.5.0
requests==2.28.2
aiohttp==3.8.4
PyYAML==6.0
SQLAlchemy==2.0.19
google-api-python-client==2.86.0
google-auth==2.18.0
feedparser==6.0.10
beautifulsoup4==4.12.2
markdown==3.4.3
pydantic==2.0.2
python-dotenv==1.0.0
Pillow==9.5.0
fastapi==0.99.1
uvicorn==0.22.0
